Cleaned up e2e test package (#24947)
no ref
- refactored PostPreview page
- removed unnecessary, descriptive comments across files
- added key press handler to remove Playwright knowledge from test
writing
- refactored Portal page handlers to the frontend page, more aligning
with use

This PR is some general housekeeping items to help unify out patterns in
the package before increasing adoption of this for our testing across
Ghost.
