+++ e2e/tests/admin/PostPreview.test.ts
@@ -23,9 +23,9 @@ test.describe('Post Preview Modal', () => {
         expect(await postEditorPage.previewModal.isVisible()).toBe(true);
 
         const postContent = await postEditorPage.previewModal.getPostContent();
-        await postContent.image().click();
+        await postContent.image.click();
 
-        await page.keyboard.press('Escape');
+        await postEditorPage.pressKey('Escape');
 
         await postEditorPage.previewModal.waitForHidden();
         expect(await postEditorPage.previewModal.isVisible()).toBe(false);
@@ -45,7 +45,7 @@ test.describe('Post Preview Modal', () => {
 
         await postEditorPage.previewModal.header.click();
 
-        await page.keyboard.press('Escape');
+        await postEditorPage.pressKey('Escape');
 
         await postEditorPage.previewModal.waitForHidden();
         expect(await postEditorPage.previewModal.isVisible()).toBe(false);
